requirements.txt
streamlit==1.22.0
google-api-python-client==2.86.0
requests==2.31.0
Pillow==9.5.0
python-dotenv==1.0.0
